Properties: Add the m5_prop/m5_user_prop structs
This commit adds two structs to handle the MQTT v5
Properties.

To integrate MQTT v5 Properties to Control Packets,
developers must declare and fill the struct with the
properties associated to the target Control Packet.

The m5_prop struct can be used for any Control Packet,
and because of this, the m5_prop struct contains all
the fields nedded to represent the properties for every
Control Packet.

Signed-off-by: Flavio Santes <flavio.santes@1byt3.com>